[FLINK-19123] Move PerJobMiniClusterJobClient to runtime.minicluster package
